Move Indent* plugins to indent/Indent.
